Cleanup and do housekeeping with plugin examples
This PR performs housekeeping of the plugin examples:

* makes the examples independent of Hive being installed
* adds "has_access" in the examples
* removes the misleading "metastore" (which is hive metastore not
  Airflow Metastore as used in other places

This way our example will be much easier to apply by anyone.
