+++ .dockerignore
@@ -34,7 +34,6 @@
 !chart
 !docs
 !licenses
-!metastore_browser
 
 # Add those folders to the context so that they are available in the CI container
 !scripts/in_container
