+++ docs/apache-airflow/empty_plugin/README.md
@@ -1,4 +1,4 @@
-{#
+<!--
  Licensed to the Apache Software Foundation (ASF) under one
  or more contributor license agreements.  See the NOTICE file
  distributed with this work for additional information
@@ -15,13 +15,16 @@
  KIND, either express or implied.  See the License for the
  specific language governing permissions and limitations
  under the License.
-#}
+-->
 
-{% extends 'metastore_browser/base.html' %}
+# Apache example plugin
 
-{% block plugin_content %}
-    <h4>
-        <span>Hive Databases</span>
-    </h4>
-    {{ table }}
-{% endblock %}
+This is an example plugin for Apache Airflow.
+
+This plugin displays empty view.
+
+## Installation
+
+To use this plugin, copy the contents of this directory to the plugins directory.
+The location of this directory is specified by  `plugins_folder` option in `core` section.
+By default, it's `{AIRFLOW_HOME}/plugins`.
