--- metastore_browser/templates/metastore_browser/base.html
@@ -1,57 +1,0 @@
-{#
- Licensed to the Apache Software Foundation (ASF) under one
- or more contributor license agreements.  See the NOTICE file
- distributed with this work for additional information
- regarding copyright ownership.  The ASF licenses this file
- to you under the Apache License, Version 2.0 (the
- "License"); you may not use this file except in compliance
- with the License.  You may obtain a copy of the License at
-
-   http://www.apache.org/licenses/LICENSE-2.0
-
- Unless required by applicable law or agreed to in writing,
- software distributed under the License is distributed on an
- "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- KIND, either express or implied.  See the License for the
- specific language governing permissions and limitations
- under the License.
-#}
-
-{% extends base_template %}
-
-{% block head %}
-  {{ super() }}
-  <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='dataTables.bootstrap.css') }}">
-  <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='select2.css') }}">
-{% endblock %}
-
-{% block body %}
-  <div>
-    <h3 style="float: left">
-        {% block page_header %}Hive Metastore Browser{% endblock%}
-    </h3>
-    <div id="object" class="select2-drop-mask" style="margin-top: 25px; width: 400px;float: right"></div>
-    <div style="clear: both"></div>
-  </div>
-  {% block plugin_content %}{% endblock %}
-{% endblock %}
-
-{% block tail %}
-  {{ super() }}
-  <script src="{{ url_for('static', filename='jquery.dataTables.min.js') }}"></script>
-  <script src="{{ url_for('static', filename='dataTables.bootstrap.js') }}"></script>
-  <script src="{{ url_for('static', filename='select2.min.js') }}"></script>
-  <script>
-    // Filling up the table selector
-    url = "{{ url_for('.objects') }}";
-    $.get(url, function( data ) {
-      $("#object").select2({
-        data: data,
-        placeholder: "Table Selector",
-      })
-      .on("change", function(e){
-        window.location = "{{ url_for('.table') }}?table=" + e.val;
-      });
-    }, "json");
-  </script>
-{% endblock %}
